chore: rename folder to common
